v24.02

+++ app/src/test/java/org/supla/android/usecases/channel/ChannelActionUseCaseTest.kt
@@ -246,7 +246,7 @@ class ChannelActionUseCaseTest {
     channelId: Int,
     channelFunc: Int,
     buttonType: ButtonType,
-    flags: Int = 0,
+    flags: Long = 0,
     actionAssertion: (ActionParameters) -> Unit
   ) {
     // given
